fix: keep session URL helpers internal
