+++ src/routing/session-key.ts
@@ -208,15 +208,6 @@ export function buildAgentMainSessionKey(params: {
   return `agent:${agentId}:${mainKey}`;
 }
 
-/** Resolves the configured main session identity for one agent and session scope. */
-export function resolveCanonicalMainSessionKey(params: {
-  agentId: string;
-  mainKey?: string | undefined;
-  sessionScope?: "per-sender" | "global";
-}): string {
-  return params.sessionScope === "global" ? "global" : buildAgentMainSessionKey(params);
-}
-
 export function buildAgentPeerSessionKey(params: {
   agentId: string;
   mainKey?: string | undefined;
